add css on popup image detail  issue #181 (#327)
* Update Account.js

* Update Account.js

* adding organization ID filter to planter's page

Adding an addition filter to search. The filter is label as the organization ID. The purpose is to search list by organization ID.

* Update FilterTopPlanter.js

Review code structure, review code implementation and functionality.

* Update Account.js

edit file to display admin username.

* Update TreeImageScrubber.js

add css on popup image detail to create more spacing for text in the pop image.